[REFACTOR] move registration to next file
