+++ resource_activity/models/__init__.py
@@ -1,4 +1,5 @@
 from . import partner
 from . import resource_activity
+from . import resource_activity_registration
 from . import sale
 from . import resource
